Remove the CONNECT event.
From now on there is no need to explicitly emit a CONNECT event to connect
different instances of scripts. The CÉU programmer should consider the scripts
are always connected. The environment should be in charge of managing the
connections based on the route table.

Known bug: sometimes a SEGFAULT happens on lua_next call (env_util.c:92) after
           receiving messages containing several Lua tables.